requirements.txt
aniso8601==2.0.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
flask-restplus==0.10.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
pprint==0.1
python-craigslist==1.0.5
pytz==2018.3
requests==2.18.4
selenium==3.9.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
